all files add to component
