+++ front-end/src/App.js
@@ -1,11 +1,11 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import './App.css';
 import Home from './Home'
-import Login from './login/Login';
-import CustomerSignUp from './signUp/CustomerSignUp';
-import ManagerSignUp from './signUp/ManagerSignUp';
-import CustomerProfile from './Profile/CustomerProfile';
-import ManagerProfile from './Profile/ManagerProfile';
+import Login from './components/login/Login';
+import CustomerSignUp from './components/signUp/CustomerSignUp';
+import ManagerSignUp from './components/signUp/ManagerSignUp';
+import CustomerProfile from './components/profile/CustomerProfile';
+import ManagerProfile from './components/profile/ManagerProfile';
 
 function App() {
   return (
