+++ front-end/src/components/login/Login.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import "../bootstrap.css"
+import "../../bootstrap.css"
 import "./Login.css"
 
 const Login = () => {
