+++ front-end/src/components/profile/EditPassword.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import '../bootstrap.css';
+import '../../bootstrap.css';
 import './EditPassword.css';
 
 function EditPassword() {
