+++ front-end/src/components/profile/SaveButton.js
@@ -1,5 +1,5 @@
 import {React, useState} from 'react';
-import '../bootstrap.css';
+import '../../bootstrap.css';
 import './CustomerProfile.css';
 
 function SaveButton() {
