+++ front-end/src/components/signUp/InputName.js
@@ -1,20 +1,17 @@
 import React, { useState } from 'react';
 import "./Input.css"
-import "../bootstrap.css"
+import "../../bootstrap.css"
 
-function InputEmail() {
-  const [email, setEmail] = useState('');
+function InputName(props) {
+  const [name, setName] = useState('');
   const [validationMessage, setValidationMessage] = useState('');
   const [showValidation, setShowValidation] = useState(false);
 
-  const validateEmail = (email) => {
-    const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    return re.test(String(email).toLowerCase());
-  };
-
-  const validateEmailButton = () => {
-    if (!validateEmail(email)) {
-      setValidationMessage('Invalid email address.');
+  const validateName = () => {
+    if (name.length > 20) {
+      setValidationMessage('Name must be less than or equal to 20 characters.');
+    } else if (name === '') {
+      setValidationMessage('Name is required.');
     } else {
       setValidationMessage('');
     }
@@ -24,11 +21,13 @@ function InputEmail() {
   return (
     <div>
       <div className="form-group">
-        <label htmlFor="email">Email</label>
+        <label htmlFor="name" className="form-label">
+          {props.business === true ? 'Business Name' : 'Name'}
+        </label>
         <div className="input-row">
           <input
-            type="email"
-            id="email"
+            type="text"
+            id="name"
             className={`form-control ${
               showValidation && validationMessage
                 ? 'is-invalid'
@@ -36,14 +35,10 @@ function InputEmail() {
                 ? 'is-valid'
                 : ''
             }`}
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
+            value={name}
+            onChange={(e) => setName(e.target.value)}
           />
-          <button
-            type="button"
-            className="btn btn-primary"
-            onClick={validateEmailButton}
-          >
+          <button type="button" className="btn btn-primary" onClick={validateName}>
             Check
           </button>
         </div>
@@ -53,4 +48,4 @@ function InputEmail() {
   );
 }
 
-export default InputEmail;
+export default InputName;
