+++ front-end/src/components/signUp/InputPassword.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import "../bootstrap.css"
+import "../../bootstrap.css"
 import "./Input.css"
 
 function InputPassword() {
