+++ front-end/src/components/signUp/ManagerSignUp.js
